add local header file to YCM path.

+++ include/VECS/comps/Transform.hpp
@@ -2,7 +2,7 @@
 
 struct Transform
 {
-	size_t x;
-	size_t y;
-	size_t z;
+	int x;
+	int y;
+	int z;
 };
